Inject RARP packets for vhostuser backend too
GUEST_ANNOUNCE will make the guest send GARPs and/or NAs (if any IP
addresses are present). But the guest will not send a RAPR.

It's important to send RARP though, for two reasons:

1. Some SDNs may expect RARP being issued, specifically. Namely, OVN
   controller with activation-strategy=rarp will expect a RARP and will
   use it as an indication that libvirt completed migration. Only then
   it will unblock the traffic for the destination port. Without RARP,
   restoring connectivity for the migrated port takes unnecessary long
   time and requires a completion action from CMS that manages OVN
   ports.

2. If virtio guest doesn't have IP addresses configured, no
   announcement frames (neither GARPs nor NAs) will be sent at all. For
   completeness, it's better to always send RARP.

There are several related requests for related projects to adjust their
behavior, e.g.:

- for kernel, to send RARP on GUEST_ANNOUNCE request:
  https://issues.redhat.com/browse/RHEL-71919

- for OVN, to handle announcement GARPs and NAs (in addition to RARP) as
  indication that migration is complete:
  https://issues.redhat.com/browse/FDP-1042

- for Neutron, to not use activation-strategy=rarp for DPDK ports:
  https://launchpad.net/bugs/2092407

(The latter can only be viewed as a temporary workaround, since there's
a good reason why activation-strategy=rarp is used - to avoid unexpected
duplicate frames delivered from both source and destination hosts due to
async nature of migration pause/unpause cycle.)

All of these may be helpful but don't address the problem completely.
This patch for qemu is an attempt to fix the scenario at the root:
namely, forcing qemu to send RARPs for all backends, regardless of what
guest virtio claims as supported. (At the end of the day, GUEST_ANNOUNCE
protocol never promised to send a RARP specifically, only that the guest
will be "announced", somehow.)

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>